refactor: 범용성을 위한 파일 이름 변경
